refactor: change folder shared to modal for better naming
