+++ src/pages/ManageWorkspaces.tsx
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { AppDispatch, RootState } from '../redux/store'
 import { useEffect, useState } from 'react'
 import { createWorkspace, getWorkspacesByUser } from '../redux/slice/workspaceSlice'
-import DialogDemo from '@/components/shared/shared/Modal'
+import DialogDemo from '@/components/shared/modal/Modal'
 import WorkspaceForm from '@/components/shared/forms/WorkspaceForm'
 import { z } from 'zod'
 import { workspaceShema } from '@/schema/modal'
