+++ src/pages/WorkspaceDetails.tsx
@@ -2,7 +2,7 @@ import Sidebar from '../components/shared/Sidebar'
 import { Download, Plus } from 'lucide-react'
 import { useLocation, Navigate, useParams, Outlet, useNavigate } from 'react-router-dom'
 import { useState, useEffect } from 'react'
-import DialogDemo from '@/components/shared/shared/Modal'
+import DialogDemo from '@/components/shared/modal/Modal'
 import InviteUserForm from '@/components/shared/forms/InviteUserForm'
 import { useSelector } from 'react-redux'
 import { selectSidebarOpen } from '../redux/features/sidebarSlice'
